Add payment methods feature and fix compilation issues: - Add PaymentMethod entity and related components - Update database schema for payment methods - Fix parameter naming in SettingsScreen - Resolve compilation errors in MainActivity and CheckbookApp - Clean up build files

+++ app/src/main/java/com/example/checkbook/ui/components/TransactionItem.kt
@@ -122,7 +122,7 @@ fun TransactionItem(
                                             color = MaterialTheme.colorScheme.onSurfaceVariant
                                         )
                                         transaction.paymentMethodId?.let { id ->
-                                            paymentMethods.find { it.id == id }?.let { paymentMethod ->
+                                            paymentMethods.find { it.id.toLong() == id }?.let { paymentMethod ->
                                                 Spacer(modifier = Modifier.height(4.dp))
                                                 Text(
                                                     text = paymentMethod.name,
